Updates to new tree sanity test
`features/rhelah/sanity_new_qcow.feature`
`features/rhelah/sanity_new_tree.feature`
`steps/rhelah.py`

The majority of this change affects the steps performed for the 'new
tree' sanity test.  Notably, it causes the system to be subscribed to
'production' and then upgraded to the latest tree available. (This makes
the assumption that the new tree that is being tested is in the 'stage'
environment.)

Additionally, the list of RPMs (and current atomic version) is collected
after upgrading to the latest 'released' tree and then after the system
has been updated to the 'stage' tree. This allows us to produce a diff
of the list for later use.

The 'new qcow' test was updated for some changes to the verbiage used in
the steps.

Finally, the necessary changes to the `steps/rhelah.py` file were made
to support the collection and retrieval of the RPM list and related
files.

Signed-off-by: Micah Abbott <miabbott@redhat.com>