rhelah: Make the rollback tests can be run together
In the old version of rollback tests assume the atomic host is a
newly started host and only have one deployment. So when we run
them together scenario 2 which try to upgrade the host will
always failed except the first one. Update the scenario to make
it can work with a host that is already upgraded.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>