requirements.txt: explicitly add jinja2
In view of the upcoming work for more in-depth testing of kubernetes, we
will requiring some templating facilities in order to easily define and
create resources. We will use Jinja2 for this.

Note that this is not a true 'added' dependency, since ansible already
brings in jinja2, but we're making it explicit here so that we are aware
of the fact that jinja2 is used directly in our code.

Unfortunately, we cannot make use of Jinja2 "through" ansible using the
template module because its use is restricted to playbooks.