kubernetes.py: account for flannel bug
There is a bug in flannel (which has been fixed, but hasn't made it to
Atomic yet) which causes the first ping to be lost, and thus causes the
assertion to fail. Let's for now switch to the requirement that we
shouldn't have 100% packet loss.

This commit can be cleanly reverted once we reach flannel v0.5.0 at the
very least (or backport the appropriate patch).