Introduce before_feature and after_feature to the framework
We have some scenarios are called before and after a set of feature, just like
some some prepare and clean up steps. And it is not related with the core of the
feature testing. So separate them from the feature files and put them into the
env_setup module. And the functions can be called based on the tags marked in
feature files.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>