Move discover host from backgroud to enviroment
Move the discover host function is used to setup attributes in
context. And this can be covered by the attribute setup based
of cfg files. So move it out from the background and it will
only run once in before_all().

And after the update only need update test.cfg to swith inventory
method bewteen dynamic and static.

Signed-off-by: Yiqiao Pu <ypu@redhat.com>