Pin 'ansible' to pre-2.0 version
With the release of Ansible 2.0, the API was entirely re-written.  What
was once a simple import of a single Ansible module has become a bit
more complex.  See the following:

http://docs.ansible.com/ansible/developing_api.html#python-api-2-0

In the short term, it is easiest to pin to 1.9.4 to keep things humming.
Longer term, we should probably look at adapting the framework to the
new API.

Signed-off-by: Micah Abbott <miabbott@redhat.com>